Start Min & Max tests by moving current tests into new class.
