+++ src/System.Linq/tests/EnumerableTests.cs
@@ -172,28 +172,6 @@ namespace System.Linq.Tests
                 "namespace",
             });
         }
-
-        [Fact]
-        public void Min()
-        {
-            var one = Enumerable.Range(1, 10).ToArray();
-            var minusTen = new [] {-1, -10, 10, 200, 1000};
-            var hundred = new [] {3000, 100, 200, 1000};
-            Assert.Equal(1, one.Min());
-            Assert.Equal(-10, minusTen.Min());
-            Assert.Equal(100, hundred.Min());
-        }
-
-        [Fact]
-        public void Max()
-        {
-            var ten = Enumerable.Range(1, 10).ToArray();
-            var minusTen = new[] {-100, -15, -50, -10};
-            var thousand = new[] {-16, 0, 50, 100, 1000};
-            Assert.Equal(10, ten.Max());
-            Assert.Equal(-10, minusTen.Max());
-            Assert.Equal(1000, thousand.Max());
-        }
     }
 }
